requirements.txt
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.6.0
python-dotenv==0.17.1
requests==2.25.1
six==1.15.0
toml==0.10.1
urllib3==1.26.4
watchdog==2.1.2
wrapt==1.12.1
xspf-lib==0.2.0
